fix vim-autopep8 command detection
With pyenv installed, autopep8 is executable, but it is just a shim with
no functionality that prints an error message. The plugin would replace
file contents with the command output without checking for return code,
so the file contents would be destroyed.

Now we execute the autopep8 command to see it it returns 0 on a
version check.